refactor: 공연 등록 버튼 PerformanceRegisterButton 컴포넌트로 추출
공연 등록 버튼의 인증 및 모달 로직을 새로운 컴포넌트로 추출하여 코드 재사용성을 향상시켰습니다.

변경 내용:
- PerformanceRegisterButton: 새 컴포넌트 생성 (인증 체크 및 모달 관리)
- Hero: 버튼 로직 제거하고 PerformanceRegisterButton으로 대체
- 불필요한 상태 및 이벤트 핸들러 제거로 코드 간소화
