+++ src/components/performance/index.ts
@@ -1,1 +1,2 @@
+export { PerformanceRegisterButton } from './performance-register-button';
 export { RegisterModal } from './register-modal';
